+++ HomeWork-6/HomeWork-6.sln
@@ -5,9 +5,9 @@ VisualStudioVersion = 16.0.30804.86
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Task2", "Task2\Task2.csproj", "{28F9CC5E-D0D7-4EE1-BB02-704C1924E6CB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Task1", "Task1-1\Task1.csproj", "{4181DD12-AA6D-47BC-856D-87BEAA304E95}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Task3", "Task3\Task3.csproj", "{089C3C63-1AEC-4EE6-A6BA-377A6A2B6CC8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Task1", "Task1\Task1.csproj", "{4181DD12-AA6D-47BC-856D-87BEAA304E95}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -19,14 +19,14 @@ Global
 		{28F9CC5E-D0D7-4EE1-BB02-704C1924E6CB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{28F9CC5E-D0D7-4EE1-BB02-704C1924E6CB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{28F9CC5E-D0D7-4EE1-BB02-704C1924E6CB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4181DD12-AA6D-47BC-856D-87BEAA304E95}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4181DD12-AA6D-47BC-856D-87BEAA304E95}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4181DD12-AA6D-47BC-856D-87BEAA304E95}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4181DD12-AA6D-47BC-856D-87BEAA304E95}.Release|Any CPU.Build.0 = Release|Any CPU
 		{089C3C63-1AEC-4EE6-A6BA-377A6A2B6CC8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{089C3C63-1AEC-4EE6-A6BA-377A6A2B6CC8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{089C3C63-1AEC-4EE6-A6BA-377A6A2B6CC8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{089C3C63-1AEC-4EE6-A6BA-377A6A2B6CC8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4181DD12-AA6D-47BC-856D-87BEAA304E95}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4181DD12-AA6D-47BC-856D-87BEAA304E95}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4181DD12-AA6D-47BC-856D-87BEAA304E95}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4181DD12-AA6D-47BC-856D-87BEAA304E95}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
